requirements.txt
aniso8601==9.0.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
filelock==3.15.1
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pip==24.0
platformdirs==4.2.2
pytz==2024.1
six==1.16.0
virtualenv==20.26.2
Werkzeug==3.0.3